requirements.txt
# LangChain & LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.20

# OpenAI
openai>=1.10.0

# Vector Store
faiss-cpu>=1.7.4
tiktoken>=0.5.0

# Web UI
streamlit>=1.30.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.0.0

# Document Processing
python-docx>=0.8.11
markdown>=3.5.0

# Web Fetching & Search
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0
